Improve round tests
